+++ custom_components/aqua_temp/common/account_types.py
@@ -1,0 +1,7 @@
+from homeassistant.backports.enum import StrEnum
+
+
+class AccountTypes(StrEnum):
+    AquaTempAndroid = ("Aqua Temp (Android)",)
+    AquaTempIOS = ("Aqua Temp (iOS)",)
+    HiTemp = "HiTemp"